+++ source/modalSubmissions/changeclubmeeting.js
@@ -2,6 +2,8 @@ const { Interaction } = require('discord.js');
 const ModalSubmission = require('../classes/ModalSubmission.js');
 const { getClubDictionary, updateClub, updateClubDetails, updateList, clubInviteBuilder, clearClubReminder, cancelClubEvent, setClubReminder, createClubEvent, scheduleClubEvent } = require('../helpers.js');
 
+const YEAR_IN_MS = 31556926000;
+
 const id = "changeclubmeeting";
 module.exports = new ModalSubmission(id,
 	/** Set the meeting time/repetition properties for the club with provided id
@@ -16,17 +18,30 @@ module.exports = new ModalSubmission(id,
 		if (fields.fields.has("nextMeeting")) {
 			const unparsedValue = fields.getTextInputValue("nextMeeting");
 			const nextMeetingInput = parseInt(unparsedValue);
-			if (nextMeetingInput) {
-				club.timeslot.setNextMeeting(nextMeetingInput);
-				clearClubReminder(club.id);
-				cancelClubEvent(club.voiceChannelId, club.timeslot.eventId, interaction.guild.scheduledEvents);
-				setClubReminder(club, interaction.guild.channels);
-				createClubEvent(club, interaction.guild);
-				if (club.timeslot.periodCount && club.isRecruiting()) {
-					scheduleClubEvent(club, interaction.guild);
-				}
+			if (!nextMeetingInput) {
+				errors.nextMeeting = `The timestamp given for the next meeting (${unparsedValue}) could not be interpreted as an integer.`;
 			} else {
-				errors["nextMeeting"] = `Could not interpret ${unparsedValue} as integer`;
+				const now = Date.now();
+				if (now > nextMeetingInput * 1000) {
+					errors.nextMeeting = `The timestamp given for the next meeting (${nextMeetingInput}) is in the past.`;
+				} else if (nextMeetingInput * 1000 > now + (5 * YEAR_IN_MS)) {
+					errors.nextMeeting = "Discord does not allow the creation of events 5 years in the future. Please schedule your next meeting later.";
+				} else {
+					club.timeslot.setNextMeeting(nextMeetingInput);
+					clearClubReminder(club.id);
+					cancelClubEvent(club, interaction.guild.scheduledEvents);
+
+					await createClubEvent(club, interaction.guild);
+					if (club.isRecruiting() && club.timeslot.periodCount) {
+						scheduleClubEvent(club, interaction.guild);
+					}
+					setClubReminder(club, interaction.guild.channels);
+				}
+			}
+
+			if (errors.nextMeeting) {
+				club.timeslot.setNextMeeting(null);
+				club.timeslot.setEventId(null);
 			}
 		}
 		if (fields.fields.has("message")) {
@@ -39,15 +54,15 @@ module.exports = new ModalSubmission(id,
 			if (periodCountInput) {
 				club.timeslot.periodCount = periodCountInput;
 			} else {
-				errors["periodCount"] = `Could not interpret ${unparsedValue} as integer`;
+				errors.periodCount = `Could not interpret ${unparsedValue} as integer`;
 			}
 		}
 		if (fields.fields.has("periodUnits")) {
 			const periodUnitsInput = fields.getTextInputValue("periodUnits");
-			if (["d", "w"].includes(periodUnitsInput)) {
+			if (["days", "weeks"].includes(periodUnitsInput)) {
 				club.timeslot.periodUnits = periodUnitsInput;
 			} else {
-				errors["periodUnits"] = `Input ${periodUnitsInput} did not match "d" (days) or "w" (weeks)`;
+				errors.periodUnits = `Input ${periodUnitsInput} did not match "days" or "weeks"`;
 			}
 		}
 		updateClubDetails(club, interaction.channel);